fix demo playback
This restores how the demos playback on DOS, not
this does not exactly mean correct. From what I can gather
version 1.2-1.3 of heretic and the mega wad used for the expansion
both ship with demos that range from kinda out of sync to very.

Tested with 1.2 shareware, which had functioning demos, and
we now fail in the exact same way that the DOS versions did
for those that are broken.